scraper/requirements.txt
# Scrapy Core
scrapy==2.11.0
scrapy-redis==0.7.3

# Zyte Stack (Enterprise Scraping)
scrapy-zyte-api==0.14.0      # Zyte API integration
zyte-smartproxy-plugin==0.6.0 # Smart Proxy Manager
scrapy-poet==0.21.0           # Page object pattern for Zyte
zyte-common-items==0.20.0     # Standard item definitions

# Browser Automation (fallback for custom needs)
playwright==1.41.0
scrapy-playwright==0.0.33

# Data Processing
beautifulsoup4==4.12.3
lxml==5.1.0
cssselect==1.2.0
parsel==1.8.1

# HTTP
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# Elasticsearch
elasticsearch==8.11.1

# Data Validation
pydantic==2.5.3
python-dateutil==2.8.2

# NLP & ML
spacy==3.7.2
simhash==2.1.2
rapidfuzz==3.6.1

# Geocoding
geopy==2.4.1

# Anti-Detection (for non-Zyte requests)
fake-useragent==1.4.0
python-dotenv==1.0.0

# Scheduling
apache-airflow==2.8.0

# Utilities
tenacity==8.2.3
python-slugify==8.0.2

# Development
pytest==7.4.4
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Shub CLI for Scrapy Cloud deployment
shub==2.15.0
